requirements_fastapi.txt
python-decouple==3.7
fastapi==0.92.0
uvicorn==0.20.0
pydantic==1.10.5